refacotoring: 컴포넌트 이름 변경
- Card 컴포넌트 -> CreditCard 컴포넌트
