+++ src/components/Card/index.tsx
@@ -1,68 +1,13 @@
+import { PropsWithChildren } from 'react';
+
 import * as S from './index.style';
-import GoogleIcon from '../GoogleIcon';
 
-import { cardCompany } from '~/constants';
-import { DefaultCard } from '~/types';
+interface InformationProps extends PropsWithChildren {}
 
-export interface CardProps extends Partial<DefaultCard> {
-  size?: 'small' | 'big';
-  onClick?: (card?: DefaultCard) => void;
-}
-
-const Card = ({
-  size = 'small',
-  company = 'none',
-  numbers = ['', '', '', ''],
-  endDate = {
-    month: 'MM',
-    day: 'YY',
-  },
-  cardUser = undefined,
-  onClick,
-}: CardProps) => {
-  const clickable = onClick ? true : false;
-  const encryptedNumbers = Array.from(numbers).map((number, index) => {
-    if (typeof number === 'string') {
-      return index < 2 ? number : '●'.repeat(number.length);
-    }
-    return number;
-  });
-  const handleClickCard = () => {
-    onClick && onClick();
-  };
-
-  return (
-    <S.Container>
-      <S.Card
-        clickable={clickable}
-        onClick={handleClickCard}
-        size={size}
-        company={company}
-      >
-        {company === 'none' ? (
-          <GoogleIcon name='add' />
-        ) : (
-          <>
-            <S.CardTop>{cardCompany[company]}</S.CardTop>
-            <S.CardMiddle>
-              <S.CardChip size={size} />
-              <S.CardNumberList>
-                {encryptedNumbers.map((number, index) => (
-                  <S.CardNumberItem key={index}>{number}</S.CardNumberItem>
-                ))}
-              </S.CardNumberList>
-            </S.CardMiddle>
-            <S.CardBottom>
-              <S.CardBottomLeft>{cardUser}</S.CardBottomLeft>
-              <S.CardBottomRight>
-                {endDate.month}/{endDate.day}
-              </S.CardBottomRight>
-            </S.CardBottom>
-          </>
-        )}
-      </S.Card>
-    </S.Container>
-  );
+const Information = ({ children }: InformationProps) => {
+  return <S.Information>{children}</S.Information>;
 };
 
-export default Card;
+export default Information;
+export { default as InformationHeader } from './InformationHeader';
+export { default as InformationBody } from './InformationBody';
