+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Button } from './Button';
 export { default as Modal } from './Modal';
 export { default as Information } from './Information';
-export { default as Card } from './Card';
+export { default as CreditCard } from './CreditCard';
 export { default as InputWrapper } from './InputWrapper';
 export { default as Input } from './Input';
 export { default as Keypad } from './Keypad';
